# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Module {
  id: number;
  title: string;
  description: string;
  level: number;
  icon: string;
  lessonCount: number;
  estimatedMinutes: number;
  createdAt: string;
}

export interface Lesson {
  id: number;
  moduleId: number;
  title: string;
  content: string;
  orderIndex: number;
  estimatedMinutes: number;
  createdAt: string;
}

export interface ModuleWithLessons {
  id: number;
  title: string;
  description: string;
  level: number;
  icon: string;
  estimatedMinutes: number;
  lessons: Lesson[];
  createdAt: string;
}

export interface Quiz {
  id: number;
  moduleId: number;
  title: string;
  description: string;
  questionCount: number;
  passingScore: number;
  createdAt: string;
}

export interface QuizQuestion {
  id: number;
  quizId: number;
  question: string;
  options: string[];
  hint?: string;
  orderIndex: number;
}

export interface QuizWithQuestions {
  id: number;
  moduleId: number;
  title: string;
  description: string;
  passingScore: number;
  questions: QuizQuestion[];
  createdAt: string;
}

export type QuizSubmissionAnswersItem = {
  questionId: number;
  selectedOption: number;
};

export interface QuizSubmission {
  userId: string;
  answers: QuizSubmissionAnswersItem[];
}

export type QuizResultAnswerFeedbackItem = {
  questionId: number;
  correct: boolean;
  correctOption: number;
  explanation?: string;
};

export type BadgeRarity = (typeof BadgeRarity)[keyof typeof BadgeRarity];

export const BadgeRarity = {
  common: "common",
  uncommon: "uncommon",
  rare: "rare",
  legendary: "legendary",
} as const;

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  rarity: BadgeRarity;
  xpValue: number;
  createdAt: string;
}

export interface QuizResult {
  quizId: number;
  userId: string;
  score: number;
  totalQuestions: number;
  passed: boolean;
  xpEarned: number;
  badgesEarned: Badge[];
  answerFeedback: QuizResultAnswerFeedbackItem[];
}

export type UserProgressQuizAttemptsItem = {
  quizId: number;
  score: number;
  passed: boolean;
  attemptedAt: string;
};

export interface UserProgress {
  userId: string;
  totalXp: number;
  level: number;
  completedLessons: number[];
  completedModules: number[];
  earnedBadges: Badge[];
  quizAttempts: UserProgressQuizAttemptsItem[];
}

export interface CompleteLessonBody {
  lessonId: number;
  moduleId: number;
}

export type DashboardSummaryRecentActivityItem = {
  type: string;
  description: string;
  timestamp: string;
};

export interface DashboardSummary {
  userId: string;
  totalXp: number;
  level: number;
  modulesCompleted: number;
  modulesTotal: number;
  lessonsCompleted: number;
  lessonsTotal: number;
  quizzesPassed: number;
  badgesEarned: number;
  currentStreak: number;
  recentActivity: DashboardSummaryRecentActivityItem[];
  nextModule?: Module;
}

export interface GlobalStats {
  totalUsers: number;
  totalLessonsCompleted: number;
  totalQuizzesTaken: number;
  totalBadgesAwarded: number;
  topBadges: Badge[];
}

export interface ForumPost {
  id: number;
  authorName: string;
  userId: string;
  title: string;
  content: string;
  replyCount: number;
  createdAt: string;
}

export interface ForumReply {
  id: number;
  postId: number;
  authorName: string;
  userId: string;
  content: string;
  createdAt: string;
}

export interface ForumPostWithReplies {
  id: number;
  authorName: string;
  userId: string;
  title: string;
  content: string;
  replies: ForumReply[];
  createdAt: string;
}

export interface CreateForumPostBody {
  authorName: string;
  userId: string;
  title: string;
  content: string;
}

export interface CreateForumReplyBody {
  authorName: string;
  userId: string;
  content: string;
}

export interface CaseStudy {
  id: number;
  slug: string;
  title: string;
  subtitle: string;
  year: number;
  category: string;
  summary: string;
  content: string;
  impact: string;
  lessons: string[];
  sources: string[];
  createdAt: string;
}

export type ListQuizzesParams = {
  moduleId?: number;
};

export type ListForumPostsParams = {
  limit?: number;
  offset?: number;
};