# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InfoShield API - Information Warfare Learning App
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules/{moduleId}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a single module with its lessons
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleWithLessons"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a single lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /progress/{userId}:
    get:
      operationId: getUserProgress
      tags: [progress]
      summary: Get user progress
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
  /progress/{userId}/complete-lesson:
    post:
      operationId: completeLesson
      tags: [progress]
      summary: Mark a lesson as complete
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteLessonBody"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
  /dashboard/{userId}:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /forum/posts:
    get:
      operationId: listForumPosts
      tags: [forum]
      summary: List forum posts
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Forum posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ForumPost"
    post:
      operationId: createForumPost
      tags: [forum]
      summary: Create a forum post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForumPostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPost"
  /forum/posts/{postId}:
    get:
      operationId: getForumPost
      tags: [forum]
      summary: Get forum post with replies
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Forum post with replies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumPostWithReplies"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /forum/posts/{postId}/replies:
    post:
      operationId: createForumReply
      tags: [forum]
      summary: Create a reply to a forum post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateForumReplyBody"
      responses:
        "201":
          description: Created reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForumReply"
...
  /case-studies/{slug}:
    get:
      operationId: getCaseStudy
      tags: [casestudies]
      summary: Get a case study by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Module:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: integer
        icon:
          type: string
        lessonCount:
          type: integer
        estimatedMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - level
        - icon
        - lessonCount
        - estimatedMinutes
        - createdAt
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        content:
          type: string
        orderIndex:
          type: integer
        estimatedMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - moduleId
        - title
        - content
        - orderIndex
        - estimatedMinutes
        - createdAt
    ModuleWithLessons:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: integer
        icon:
          type: string
        estimatedMinutes:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - level
        - icon
        - estimatedMinutes
        - lessons
        - createdAt
...
    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, legendary]
        xpValue:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - icon
        - rarity
        - xpValue
        - createdAt
    UserProgress:
      type: object
      properties:
        userId:
          type: string
        totalXp:
          type: integer
        level:
          type: integer
        completedLessons:
          type: array
          items:
            type: integer
        completedModules:
          type: array
          items:
            type: integer
        earnedBadges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        quizAttempts:
          type: array
          items:
            type: object
            properties:
              quizId:
                type: integer
              score:
                type: integer
              passed:
                type: boolean
              attemptedAt:
                type: string
                format: date-time
            required:
              - quizId
              - score
              - passed
              - attemptedAt
      required:
        - userId
        - totalXp
        - level
        - completedLessons
        - completedModules
        - earnedBadges
        - quizAttempts
    CompleteLessonBody:
      type: object
      properties:
        lessonId:
          type: integer
        moduleId:
          type: integer
      required:
        - lessonId
        - moduleId
    DashboardSummary:
      type: object
      properties:
        userId:
          type: string
        totalXp:
          type: integer
        level:
          type: integer
        modulesCompleted:
          type: integer
        modulesTotal:
          type: integer
        lessonsCompleted:
          type: integer
        lessonsTotal:
          type: integer
        quizzesPassed:
          type: integer
        badgesEarned:
          type: integer
        currentStreak:
          type: integer
        recentActivity:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
              timestamp:
                type: string
                format: date-time
            required:
              - type
              - description
              - timestamp
        nextModule:
          $ref: "#/components/schemas/Module"
      required:
        - userId
        - totalXp
        - level
        - modulesCompleted
        - modulesTotal
        - lessonsCompleted
        - lessonsTotal
        - quizzesPassed
        - badgesEarned
        - currentStreak
        - recentActivity
...
    ForumPost:
      type: object
      properties:
        id:
          type: integer
        authorName:
          type: string
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        replyCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorName
        - userId
        - title
        - content
        - replyCount
        - createdAt
    ForumReply:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorName:
          type: string
        userId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorName
        - userId
        - content
        - createdAt
    ForumPostWithReplies:
      type: object
      properties:
        id:
          type: integer
        authorName:
          type: string
        userId:
          type: string
        title:
          type: string
        content:
          type: string
        replies:
          type: array
          items:
            $ref: "#/components/schemas/ForumReply"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorName
        - userId
        - title
        - content
        - replies
        - createdAt
    CreateForumPostBody:
      type: object
      properties:
        authorName:
          type: string
        userId:
          type: string
        title:
          type: string
        content:
          type: string
      required:
        - authorName
        - userId
        - title
        - content
    CreateForumReplyBody:
      type: object
      properties:
        authorName:
          type: string
        userId:
          type: string
        content:
          type: string
      required:
        - authorName
        - userId
        - content
    CaseStudy:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
        year:
          type: integer
        category:
          type: string
        summary:
          type: string
        content:
          type: string
        impact:
          type: string
        lessons:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - subtitle
        - year
        - category
        - summary
        - content
        - impact
        - lessons
        - sources
        - createdAt