# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRecentActivityItem } from "./dashboardSummaryRecentActivityItem";
import type { Module } from "./module";

export interface DashboardSummary {
  userId: string;
  totalXp: number;
  level: number;
  modulesCompleted: number;
  modulesTotal: number;
  lessonsCompleted: number;
  lessonsTotal: number;
  quizzesPassed: number;
  badgesEarned: number;
  currentStreak: number;
  recentActivity: DashboardSummaryRecentActivityItem[];
  nextModule?: Module;
}