# lib/api-zod/src/generated/types/forumPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */

export interface ForumPost {
  id: number;
  authorName: string;
  userId: string;
  title: string;
  content: string;
  replyCount: number;
  createdAt: Date;
}