# lib/api-zod/src/generated/types/forumPostWithReplies.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */
import type { ForumReply } from "./forumReply";

export interface ForumPostWithReplies {
  id: number;
  authorName: string;
  userId: string;
  title: string;
  content: string;
  replies: ForumReply[];
  createdAt: Date;
}