# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */

export interface Module {
  id: number;
  title: string;
  description: string;
  level: number;
  icon: string;
  lessonCount: number;
  estimatedMinutes: number;
  createdAt: Date;
}