# lib/api-zod/src/generated/types/userProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoShield API - Information Warfare Learning App
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from "./badge";
import type { UserProgressQuizAttemptsItem } from "./userProgressQuizAttemptsItem";

export interface UserProgress {
  userId: string;
  totalXp: number;
  level: number;
  completedLessons: number[];
  completedModules: number[];
  earnedBadges: Badge[];
  quizAttempts: UserProgressQuizAttemptsItem[];
}